Adding the .gitignore file with an exception for node_modules
As we plan to start using webpack in the plugin project to build the js files
locally, we will install the npm repository. This change will prevent th node_modules
from getting into git repository, as those files are temporal and should not
be tracked.

This patch doesnt have follow ups, because I wanted to test the newly discussed
update process.