Update js deps and webpack config
Webpack and html loader have removed the inline "style" attributes text from mustache templates, which caused a broken appearance.
Im not sure exactly why it was caused, so in order to find a solution I have updated the deps, removed everything I could remove from package.json
and updated the options of html-loader. This seems to have solved the problem.

Ref: S5P-120