Refactor payment method name whitelist
The current implementation of payment method whitelisting relies on default naming (incl. case-sensitivity) of the methods, which may cause issues when the default name is change.
Also, only a small set of default payment methods was included, which is now solved via a text field to whitelist payment methods specifically.