Refactor a bit
I've added the missing whitelist for paypal express checkout customers, because its important and we discussed it. I also restructed the implementation methods, to make
"checkAdressesOrOpenModals" method a bit more readable. The normalization used for comparison is hidden now, because its not a detail
that is important to understand the conditions in the if block.

I've also added some explanatory comments with some phpdoc types, because they make the code easier to understand for humans and for
statical code analysis.

Ref: S5P-126