Add the 5 meta and session attributes
In order to integrate the 2. version of js-sdk we need a way to save, read and display some meta and session data.
The easiest solution seems to be the attribute system in shopware 5, which would allow us to extend the address and profile tables
with additional columns (attribute system).
Unlucky the attribute system wasnt implemented for customer profile so additionaly to attributes we need to extend the
account and register controller to load, display and save profile specific attributes.

In this commit we add the attributes, make sure they have individual and translatable names and extend necessary controller
so that the attributes can be read and saved.

The solution is tested with 5.7 and 5.3 versions of shopware 5 to ensure we are not missing anything. The build script is midified to
correctly create the store version, that has slightly changed names of the attributes (to prevent collision)

Ref: S5P-124